Grin.Exp

Documentation

data BPat

Bind patterns that can be found in the EBind epxressions.

Constructors

BNodePat Tag [Name]

BVar Name

BUnit

data CPat

Case patterns that can be found in the Alt expressions.

Constructors

NodePat Tag [Name]

LitPat SimpleValue

DefaultPat

data External

Constructors

External

Fields

eName :: Name

eRetType :: Ty

eArgsType :: [Ty]

eEffectful :: Bool

data Exp

Constructors

Program [External] [Def]

Def Name [Name] Exp

EBind Exp BPat Exp

ECase Name [Alt]

SApp Name [Name]

SPure VarOrValue

SStore Name

SFetch Name

SUpdate Name Name

Alt CPat Exp

Block Exp

type Program = Exp

type Def = Exp

type Alt = Exp

type SimpleExp = Exp

isExternalName :: [External] -> Name -> Bool

externals :: Exp -> [External]

programToDefs :: Program -> Map Name Exp

data RenderingOption

Constructors

Simple

WithExternals

data ExpF r

Constructors

ProgramF [External] [r]

DefF Name [Name] r

EBindF r BPat r

ECaseF Name [r]

SAppF Name [Name]

SPureF VarOrValue

SStoreF Name

SFetchF Name

SUpdateF Name Name

AltF CPat r

BlockF r

prettyExternals :: [External] -> Doc

prettyProgram :: RenderingOption -> Exp -> Doc

prettyHighlightExternals :: [External] -> Exp -> Doc

Print a given expression with highlighted external functions.

printGrin :: Exp -> IO ()

Orphan instances

Pretty Ty

Instance details

Methods

pretty :: Ty -> Doc

prettyList :: [Ty] -> Doc