Grin.GExp

Documentation

type family Elem (c :: ExpCtx) (xs :: [ExpCtx]) (cs :: [ExpCtx]) :: Constraint where ...

Equations

Elem c xs (c ': _) = ()

Elem c xs (d ': cs) = Elem c xs cs

Elem c xs '[] = TypeError ((Text "Expected expression type " :<>: ShowType xs) :$$: (Text "but got " :<>: ShowType c))

type IsExp c cs = Elem c cs cs

data Exp (ctx :: ExpCtx) where

Constructors

Program

Fields

:: [External]

-> [Exp Def_]

definitions

-> Exp Prg

Def

Fields

:: Name

-> [Name]

arguments

-> Exp Bind_

-> Exp Def_

App

Fields

:: Name

-> [Name]

arguments

-> Exp Simple

Pure :: VarOrValue -> Exp Simple

Store :: Name -> Exp Simple

Fetch :: Name -> Exp Simple

Update

Fields

:: Name

reference to update

-> Name

new value

-> Exp Simple

Alt

Fields

:: CPat

-> Exp Bind_

continuation

-> Exp Alt_

Case

Fields

:: Name

scrutinee

-> [Exp Alt_]

possible alternatives of a case

-> Exp Case_

Bind :: (IsExp lhs [Simple, Case_], IsExp rhs [Simple, Case_, Bind_]) => Exp lhs -> BPat -> Exp rhs -> Exp Bind_
Ebind lhs bpat rhs

corresponds to > lhs >>= bpat -> rhs

data ExpCtx

Constructors

Simple

Bind_

Case_

Alt_

Def_

Prg

data BPat

Bind patterns that can be found in the EBind epxressions.

Constructors

BNodePat Tag [Name]

BVar Name

BUnit

data CPat

Case patterns that can be found in the Alt expressions.

Constructors

NodePat Tag [Name]

LitPat SimpleValue

DefaultPat

data External

Constructors

External

Fields

eName :: Name

eRetType :: Ty

eArgsType :: [Ty]

eEffectful :: Bool

module Grin.Value

module Grin.TypeEnv