Grin.Interpreter.Abstract.Base

Documentation

data Loc

Constructors

Loc Name

data ST

Simple type for simple values

Constructors

ST_Int64

ST_Word64

ST_Float

ST_Bool

ST_Char

ST_Loc Loc

data Node

Node type consist of a Tag and a list with simple types

Constructors

Node Tag [ST]

data T

The final type could be a Simple-type a Node type or a Unit type

Constructors

ST ST

NT Node

UT

typeOfSimpleValue :: SimpleValue -> T

data AbsEnv

The Abstract environment consist of externals, functions defined in the program and an Environment which can vary.

Constructors

AbsEnv

Fields

_absExt :: Map Name (T, [T])

_absEnv :: Env T

_absFun :: Map Name Exp

absFun :: Lens' AbsEnv (Map Name Exp)

absExt :: Lens' AbsEnv (Map Name (T, [T]))

absEnv :: Lens' AbsEnv (Env T)

newtype FunctionT

The function type collects the set of possible types.

Constructors

FunctionT (Set T, [Set T])

type AbsStore = Store Loc (Set Node)

The abstract store associates a set of possible node types with the abstract location.

data TypeEnv

The type-env that is computed during the abstract interpretation.

Constructors

TypeEnv

Fields

_heap :: AbsStore

_variable :: Env (Set T)

_function :: Map Name FunctionT

newtype AbsState

The abstract state consists only the Abstract Store.

Constructors

AbsState

Fields

_absStr :: AbsStore

absStr :: Lens' AbsState AbsStore

newtype AbstractT m a

Constructors

AbstractT

Fields

abstractT :: RWST AbsEnv () AbsState (LogicT (RWST (TypeEnv, Cache) () (TypeEnv, Cache) m)) a

runAbstractT :: (Monad m, MonadFail m, MonadIO m) => Program -> [(Name, (T, [T]))] -> AbstractT m a -> m ([(a, AbsState, ())], (TypeEnv, Cache), ())

How to run the AbstractT monad-stack

forMonadPlus :: MonadPlus m => [a] -> (a -> m b) -> m b

Simple operator to handle branching.

data CExp

Constructors

CApp Name [Name]

data Config

The visited configuration

Constructors

Config

Fields

cfgExp :: CExp

cfgStore :: AbsStore

data Cache

The cache associates a function call with a set of the type of the return value and the state of the abstract store.

Constructors

Cache (Map Config (Set (T, AbsStore)))

Cache operations

inCache :: Config -> Cache -> Bool

getCache :: Config -> Cache -> [(T, AbsStore)]

insertCache :: Config -> [(T, AbsStore)] -> Cache -> Cache

Monoid instances