Grin.Interpreter.Abstract.Interpreter

Documentation

getCacheOut :: Monad m => AbstractT m Cache

putCacheOut :: Monad m => Cache -> AbstractT m ()

updateCacheOut :: Monad m => (Cache -> Cache) -> AbstractT m ()

collectEnv :: Monad m => Env T -> AbstractT m ()

Joins the given Env T to the end result

collectFunctionType :: Monad m => (Name, [T], T) -> AbstractT m ()

Joins the type of the funcion to the end result

askCacheIn :: Monad m => AbstractT m Cache

localCacheIn :: Monad m => Cache -> AbstractT m a -> AbstractT m a

cache2TypeEnv :: Cache -> TypeEnv

exp2CExp :: Exp -> Maybe CExp

Fixpoint finding algorithm

putStore :: Store Loc (Set Node) -> AbstractT m ()

getStore :: AbstractT m (Store Loc (Set Node))

evalCache :: (Monad m, MonadFail m, MonadIO m) => ((Exp -> AbstractT m T) -> Exp -> AbstractT m T) -> (Exp -> AbstractT m T) -> Exp -> AbstractT m T

mlfp :: (MonadIO m, Monad m) => (Cache -> AbstractT m Cache) -> AbstractT m Cache

fixCache :: (MonadFail m, MonadIO m) => (t -> AbstractT m a) -> t -> AbstractT m ()

Orphan instances

(Monad m, MonadIO m, MonadFail m) => Interpreter (AbstractT m)

Instance details

Associated Types

type Val (AbstractT m) :: Type

type HeapVal (AbstractT m) :: Type

type Addr (AbstractT m) :: Type

Methods

value :: Value -> AbstractT m (Val (AbstractT m))

val2addr :: Val (AbstractT m) -> AbstractT m (Addr (AbstractT m))

addr2val :: Addr (AbstractT m) -> AbstractT m (Val (AbstractT m))

heapVal2val :: HeapVal (AbstractT m) -> AbstractT m (Val (AbstractT m))

val2heapVal :: Val (AbstractT m) -> AbstractT m (HeapVal (AbstractT m))

unit :: AbstractT m (Val (AbstractT m))

bindPattern :: Val (AbstractT m) -> (Tag, [Name]) -> AbstractT m [(Name, Val (AbstractT m))]

askEnv :: AbstractT m (Env (Val (AbstractT m)))

localEnv :: Env (Val (AbstractT m)) -> AbstractT m (Val (AbstractT m)) -> AbstractT m (Val (AbstractT m))

lookupFun :: Name -> AbstractT m Exp

isExternal :: Name -> AbstractT m Bool

external :: Name -> [Val (AbstractT m)] -> AbstractT m (Val (AbstractT m))

evalCase :: (Exp -> AbstractT m (Val (AbstractT m))) -> Val (AbstractT m) -> [Alt] -> AbstractT m (Val (AbstractT m))

funCall :: (Exp -> AbstractT m (Val (AbstractT m))) -> Name -> [Val (AbstractT m)] -> AbstractT m (Val (AbstractT m))

allocStore :: Name -> AbstractT m (Val (AbstractT m))

fetchStore :: Val (AbstractT m) -> AbstractT m (Val (AbstractT m))

extStore :: Val (AbstractT m) -> Val (AbstractT m) -> AbstractT m ()