Tutorial.Chapter01.Exercise02

Documentation

data Value

Constructors

Prim SValue

Node Node

Unit

data Node

Constructors

N

Fields

tag :: Tag

args :: [SValue]

data SValue

Constructors

SInt64 Int64

SWord64 Word64

SFloat Float

SBool Bool

SChar Char

SLoc Address

type Address = Int

For simplicity's sake, we will represent addresses using Ints.

type InterpretExternal = Map Name ([Value] -> IO Value)

How to interpret External names in the interpreter.

data Functions

Constructors

Functions

Fields

functions :: Map Name Exp

Functions defined within the GRIN program

externals :: InterpretExternal

Externals used within the GRIN program

newtype Definitional m a

Constructors

Definitional (RWST (Functions, Env Value) () (Store Address Node) m a)

type DC m = (Monad m, MonadIO m, MonadFail m)

Collection of the needed constraints.

interpreter :: InterpretExternal -> Program -> IO Value

Implementation details

simpleValue :: SimpleValue -> SValue

Turns a Simple Value from a syntax to a simple value of the semantics.

valueOf :: DC m => Name -> Definitional m Value

Looks up a name from the active frame/environment and returns its value.

svalueOf :: DC m => Name -> Definitional m SValue

Looks up a name from the active frame/environment and returns its value, expecting a simple value.

alloc :: DC m => Definitional m Address

Creates a new location that can be used in the Store operation. The size of the underlying Map will be always the last created location +1, which serves the purpose of a new address.

eval :: DC m => Exp -> Definitional m Value

value :: DC m => Value -> Definitional m Value

How to turn a source defined value to a runtime value

val2addr :: DC m => Value -> Definitional m Address

Convert a runtime value to an address value

addr2val :: DC m => Address -> Definitional m Value

Convert an address value to a runtime value.

heapVal2val :: DC m => Node -> Definitional m Value

Convert a heap value, which is a node to a runtime value

val2heapVal :: DC m => Value -> Definitional m Node

Convert a runtime value to a Node value

unit :: DC m => Definitional m Value

Creates the Unit value, which is only created when the Update operation runs.

bindPattern :: DC m => Value -> (Tag, [Name]) -> Definitional m [(Name, Value)]

Creates a list of Name and runtime value pairs which extends the environment for the right hand side of the bind. See in lazyAdd or in sumSimple

askEnv :: DC m => Definitional m (Env Value)

Return the environment, which associates names with values

localEnv :: DC m => Env Value -> Definitional m Value -> Definitional m Value

Sets the environment to the given one, this is for binds, function calls, and alternatives.

lookupFun :: DC m => Name -> Definitional m Exp

Lookup a function by its name. It should return the Def constructor which contains the parameters and the body of the function.

isExternal :: DC m => Name -> Definitional m Bool

Checks if the given name refers to an external function.

external :: DC m => Name -> [Value] -> Definitional m Value

Run the given external with the parameters

funCall :: DC m => (Exp -> Definitional m Value) -> Name -> [Value] -> Definitional m Value

evalCase :: DC m => (Exp -> Definitional m Value) -> Value -> [Alt] -> Definitional m Value

allocStore :: DC m => Name -> Definitional m Value

Creates a location for a given name. This is particular for the GRIN store structure, where the Store operation must be part of a Bind, thus there will be always a name to bind to, which should hold the address of the created location.

In this Definitional interpreter for every Store operation that the interpreter evaluates it must creates a new location.

fetchStore :: DC m => Value -> Definitional m Value

Loads the content from the store addressed by the given value

extStore :: DC m => Value -> Value -> Definitional m ()

Extends the store with the given value.

The externals that the interpreter can understand

knownExternals :: Map Name ([Value] -> IO Value)