Tutorial.Chapter02.Exercise01

Documentation

eval :: (MonadIO m, Interpreter m) => (Exp -> m (Val m)) -> Exp -> m (Val m)

The interpreter is written in an open recursive style; the evaluator does not call itself recursively, instead it takes as an argument a function ev (the argument) is called instead of self-recursion. This is a standard encoding for recursive functions in a setting without recursive binding. It is up to an external function, such as the Y-combinator, to close the recursive loop. This open recursive form is crucial because it allows intercepting recursive calls to perform “deep” instrumentation of the interpreter.

class (Monad m, MonadFail m) => Interpreter m where

Associated Types

type Val m :: *

type HeapVal m :: *

type Addr m :: *

Methods

value :: Value -> m (Val m)

val2addr :: Val m -> m (Addr m)

addr2val :: Addr m -> m (Val m)

heapVal2val :: HeapVal m -> m (Val m)

val2heapVal :: Val m -> m (HeapVal m)

unit :: m (Val m)

bindPattern :: Val m -> (Tag, [Name]) -> m [(Name, Val m)]

askEnv :: m (Env (Val m))

Return the computational environment

localEnv :: Env (Val m) -> m (Val m) -> m (Val m)

Set the local environment

lookupFun :: Name -> m Exp

isExternal :: Name -> m Bool

external :: Name -> [Val m] -> m (Val m)

evalCase :: (Exp -> m (Val m)) -> Val m -> [Alt] -> m (Val m)

funCall :: (Exp -> m (Val m)) -> Name -> [Val m] -> m (Val m)

allocStore :: Name -> m (Val m)

fetchStore :: Val m -> m (Val m)

extStore :: Val m -> Val m -> m ()