Tutorial.Chapter03.Exercise02

Documentation

sparseCaseOptimisation :: TypeEnv -> Exp -> Exp

Sparse case optimisation uses ana to transform the program from top to down. It checks if the given alterntive can be removed. The alternative can be removed if it does not match any of the Nodes in the type associated with the variable. Literal matching alternatives must be kept.

matchingAlt :: Type -> CPat -> Bool

Returns True if the given pattern can be matched with the type of the scrutinee

removeTheRedundantDefault :: Type -> [Alt] -> [Alt]

Remove the redundant detault